chore: move some dirs around
