+++ dimos/agents/testing/vlm_stream_tester.py
