+++ dimos/core/demos/stress_test_module.py
