+++ dimos/core/test_mcp_integration.py
@@ -38,13 +38,13 @@ from dimos.agents.mcp.mcp_adapter import McpAdapter
 from dimos.agents.mcp.mcp_server import McpServer
 from dimos.core.coordination.blueprints import autoconnect
 from dimos.core.coordination.module_coordinator import ModuleCoordinator
+from dimos.core.demos.stress_test_module import StressTestModule
 from dimos.core.global_config import global_config
 from dimos.core.run_registry import (
     RunEntry,
     cleanup_stale,
     list_runs,
 )
-from dimos.core.tests.stress_test_module import StressTestModule
 from dimos.robot.cli.dimos import main
