+++ dimos/imitation/collection/recorder.py
@@ -26,7 +26,7 @@ and `coordinator_joint_state` (observation), `status` (episode segmentation).
 from __future__ import annotations
 
 from dimos.core.stream import In
-from dimos.learning.collection.episode_monitor import EpisodeStatus
+from dimos.imitation.collection.episode_monitor import EpisodeStatus
 from dimos.memory2.module import Recorder, RecorderConfig
 from dimos.msgs.sensor_msgs.Image import Image
 from dimos.msgs.sensor_msgs.JointState import JointState
