+++ dimos/imitation/dataprep/formats/lerobot/writer.py
@@ -46,8 +46,8 @@ from typing import Any
 import numpy as np
 from numpy.typing import NDArray
 
-from dimos.learning.dataprep.core import DEFAULT_FPS, OutputConfig, Sample, is_image_array
-from dimos.learning.dataprep.formats._stats import StreamingStats, stats_from_metadata
+from dimos.imitation.dataprep.core import DEFAULT_FPS, OutputConfig, Sample, is_image_array
+from dimos.imitation.dataprep.formats._stats import StreamingStats, stats_from_metadata
 from dimos.utils.logging_config import setup_logger
 
 logger = setup_logger()
