+++ dimos/imitation/dataprep/formats/test_lerobot.py
@@ -31,9 +31,9 @@ import pandas as pd
 import pyarrow.parquet as pq
 import pytest
 
-from dimos.learning.dataprep.core import OutputConfig, Sample
-from dimos.learning.dataprep.formats.lerobot.reader import inspect
-from dimos.learning.dataprep.formats.lerobot.writer import write
+from dimos.imitation.dataprep.core import OutputConfig, Sample
+from dimos.imitation.dataprep.formats.lerobot.reader import inspect
+from dimos.imitation.dataprep.formats.lerobot.writer import write
 
 
 def _state_samples(n: int = 4) -> Iterator[Sample]:
