+++ dimos/porcelain/test_remote_module_source.py
@@ -21,9 +21,9 @@ import pytest
 
 from dimos.core.coordination.module_coordinator import ModuleCoordinator
 from dimos.core.core import rpc
+from dimos.core.demos.stress_test_module import StressTestModule
 from dimos.core.global_config import GlobalConfig
 from dimos.core.module import Module
-from dimos.core.tests.stress_test_module import StressTestModule
 from dimos.porcelain.dimos import Dimos
 from dimos.porcelain.remote_module_source import RemoteModuleSource, _RemoteProxy
