+++ pyproject.toml
@@ -206,7 +206,7 @@ visualization = [
 ]
 
 learning = [
-    # dimos.learning.dataprep dataset writers (lazy-imported per format)
+    # dimos.imitation.dataprep dataset writers (lazy-imported per format)
     "pyarrow",  # LeRobot v3.0 data/episodes parquet
     "pandas",   # LeRobot v3.0 tasks.parquet (task-indexed)
     "h5py",     # HDF5 writer
